Refactor Institutions.jsx with accounts helpers.
